move NettyUtils to common
