+++ common/src/main/java/tachyon/network/NettyUtils.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.netty;
+package tachyon.network;
 
 import java.util.concurrent.ThreadFactory;
 
@@ -25,6 +25,7 @@ import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
 
 import tachyon.util.ThreadFactoryUtils;
+import tachyon.worker.netty.ChannelType;
 
 public final class NettyUtils {
   private NettyUtils() {}
